+++ src/spice2x/hooks/graphics/backends/d3d9/d3d9_screenshot.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <d3d9.h>
+
+void graphics_d3d9_process_capture(
+	IDirect3DDevice9 *device,
+	IDirect3DSwapChain9 *sub_swap_chain);
